Revert "Add command signal handler. (#2685)" (#2691)
Reason for removing - this signal handler addresses only a single kind of clean-up, but I had a more generalized and adaptive implementation in mind.  If we do it this way, I can imagine more functionality being added into this and it becomes a "kitchen sink" trying to handle more and more different types of situations as they arise.. it conflates many concerns, and the more responsibilities it gains, the harder it gets to refactor it away.

What I envisioned is the signal handler setting a flag, or firing a hook, that other parts of terminus's code detect and do their own clean-up when it happens.. so if I'm terminus command X, I decide what needs to happen if I'm interrupted.

If what I imagined is infeasible, it's not hard to revert this revert and put this signal handler back, but I'd like to pause on putting it in right away.